[#547](refact) Remove view as argument
